Refactor code structure - All header file now sits within include/ directory - Remove Common.hh - All files now individually include required standard libraries
