+++ common/event-queue/Event.cc
@@ -3,12 +3,15 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "common/event-queue/Event.hh"
+#include "common/Event.hh"
+#include <cassert>
 
 using namespace NetworkAnalytical;
 
-Event::Event(Callback callback, CallbackArg callback_arg) noexcept
-    : callback(callback), callback_arg(callback_arg) {}
+Event::Event(const Callback callback, const CallbackArg callback_arg) noexcept
+    : callback(callback), callback_arg(callback_arg) {
+  assert(callback != nullptr);
+}
 
 void Event::invoke_event() noexcept {
   // check the validity of the event
