+++ common/event-queue/EventList.cc
@@ -3,11 +3,15 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "common/event-queue/EventList.hh"
+#include "common/EventList.hh"
+#include <cassert>
 
 using namespace NetworkAnalytical;
 
-EventList::EventList(EventTime event_time) noexcept : event_time(event_time) {
+EventList::EventList(const EventTime event_time) noexcept
+    : event_time(event_time) {
+  assert(event_time >= 0);
+
   // create an empty event list
   events = std::list<Event>();
 }
@@ -17,8 +21,10 @@ EventTime EventList::get_event_time() const noexcept {
 }
 
 void EventList::add_event(
-    Callback callback,
-    CallbackArg callback_arg) noexcept {
+    const Callback callback,
+    const CallbackArg callback_arg) noexcept {
+  assert(callback != nullptr);
+
   // add the event to the event list
   events.emplace_back(callback, callback_arg);
 }
