+++ congestion_aware/basic-topology/FullyConnected.cc
@@ -3,7 +3,8 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "congestion_aware/basic-topology/FullyConnected.hh"
+#include "congestion_aware/FullyConnected.hh"
+#include <cassert>
 
 using namespace NetworkAnalyticalCongestionAware;
