+++ congestion_aware/example.cc
@@ -3,11 +3,11 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "common/Common.hh"
-#include "common/event-queue/EventQueue.hh"
-#include "common/network-parser/NetworkParser.hh"
-#include "congestion_aware/network/Chunk.hh"
-#include "congestion_aware/topology/Helper.hh"
+#include <iostream>
+#include "common/EventQueue.hh"
+#include "common/NetworkParser.hh"
+#include "congestion_aware/Chunk.hh"
+#include "congestion_aware/Helper.hh"
 
 using namespace NetworkAnalytical;
 using namespace NetworkAnalyticalCongestionAware;
