+++ congestion_unaware/basic-topology/Ring.cc
@@ -3,7 +3,8 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "congestion_unaware/basic-topology/Ring.hh"
+#include "congestion_unaware/Ring.hh"
+#include <cassert>
 
 using namespace NetworkAnalytical;
 using namespace NetworkAnalyticalCongestionUnaware;
