+++ congestion_unaware/example.cc
@@ -3,9 +3,9 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "common/Common.hh"
-#include "common/network-parser/NetworkParser.hh"
-#include "congestion_unaware/topology/Helper.hh"
+#include <iostream>
+#include "common/NetworkParser.hh"
+#include "congestion_unaware/Helper.hh"
 
 using namespace NetworkAnalytical;
 using namespace NetworkAnalyticalCongestionUnaware;
