+++ congestion_unaware/topology/Topology.cc
@@ -3,16 +3,29 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 *******************************************************************************/
 
-#include "congestion_unaware/topology/Topology.hh"
+#include "congestion_unaware/Topology.hh"
+#include <cassert>
 
 using namespace NetworkAnalytical;
 using namespace NetworkAnalyticalCongestionUnaware;
 
-Topology::Topology() noexcept : npus_count(-1) {}
+Topology::Topology() noexcept : npus_count(-1), dims_count(-1) {}
 
 int Topology::get_npus_count() const noexcept {
   assert(npus_count > 0);
 
   // return npus_count
   return npus_count;
+}
+
+int Topology::get_dims_count() const noexcept {
+  assert(dims_count > 0);
+
+  return dims_count;
+}
+
+std::vector<int> Topology::get_npus_count_per_dim() const noexcept {
+  assert(npus_count_per_dim.size() == dims_count);
+
+  return npus_count_per_dim;
 }
