+++ include/astra-network-analytical/common/Event.hh
@@ -5,7 +5,6 @@ LICENSE file in the root directory of this source tree.
 
 #pragma once
 
-#include "common/Common.hh"
 #include "common/Type.hh"
 
 namespace NetworkAnalytical {
