+++ include/astra-network-analytical/congestion_aware/Ring.hh
@@ -5,27 +5,41 @@ LICENSE file in the root directory of this source tree.
 
 #pragma once
 
-#include "common/Common.hh"
 #include "common/Type.hh"
-#include "congestion_unaware/basic-topology/BasicTopology.hh"
+#include "congestion_aware/BasicTopology.hh"
 
 using namespace NetworkAnalytical;
 
-namespace NetworkAnalyticalCongestionUnaware {
+namespace NetworkAnalyticalCongestionAware {
 
+/**
+ * A ring basic-topology.
+ */
 class Ring final : public BasicTopology {
  public:
-  explicit Ring(
+  /**
+   * Constructor.
+   *
+   * @param npus_count number of npus in a ring
+   * @param bandwidth bandwidth of link
+   * @param latency latency of link
+   * @param bidirectional true if ring is bidirectional, false otherwise
+   */
+  Ring(
       int npus_count,
       Bandwidth bandwidth,
       Latency latency,
       bool bidirectional = true) noexcept;
 
-  [[nodiscard]] int compute_hops_count(DeviceId src, DeviceId dest)
+  /**
+   * Implementation of route function.
+   */
+  [[nodiscard]] Route route(DeviceId src, DeviceId dest)
       const noexcept override;
 
  private:
+  /// mark whether the ring is bidirectional.
   bool bidirectional;
 };
 
-} // namespace NetworkAnalyticalCongestionUnaware
+} // namespace NetworkAnalyticalCongestionAware
