+++ include/astra-network-analytical/congestion_unaware/Ring.hh
@@ -5,20 +5,26 @@ LICENSE file in the root directory of this source tree.
 
 #pragma once
 
-#include "common/Common.hh"
 #include "common/Type.hh"
-#include "congestion_unaware/basic-topology/BasicTopology.hh"
+#include "congestion_unaware/BasicTopology.hh"
 
 using namespace NetworkAnalytical;
 
 namespace NetworkAnalyticalCongestionUnaware {
 
-class FullyConnected final : public BasicTopology {
+class Ring final : public BasicTopology {
  public:
-  FullyConnected(int npus_count, Bandwidth bandwidth, Latency latency) noexcept;
+  explicit Ring(
+      int npus_count,
+      Bandwidth bandwidth,
+      Latency latency,
+      bool bidirectional = true) noexcept;
 
   [[nodiscard]] int compute_hops_count(DeviceId src, DeviceId dest)
       const noexcept override;
+
+ private:
+  bool bidirectional;
 };
 
 } // namespace NetworkAnalyticalCongestionUnaware
